# all: Temporarily revert to BurntSushi for TOML front matter handling