We still have go-toml as a transitive dependency, and it is the way to go eventually, but we care about speed, so let us wait that one out.

Note that the issue this fixes is about taxonomies, but I guess this is a general issue for sites with many pages that uses TOML as front matter.

```
benchmark                                                                                                                      old ns/op     new ns/op     delta
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=0,shortcodes=false,render=false-4      559588018     519948664     -7.08%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=1,shortcodes=false,render=false-4      777332160     517742822     -33.39%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=false,render=false-4      731425010     579082320     -20.83%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=20,shortcodes=false,render=false-4     955919053     680114689     -28.85%

benchmark                                                                                                                      old allocs     new allocs     delta
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=0,shortcodes=false,render=false-4      3538011        2967864        -16.11%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=1,shortcodes=false,render=false-4      3755649        3052225        -18.73%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=false,render=false-4      4783132        3343569        -30.10%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=20,shortcodes=false,render=false-4     8749762        4352321        -50.26%

benchmark                                                                                                                      old bytes     new bytes     delta
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=0,shortcodes=false,render=false-4      371997008     340437445     -8.48%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=1,shortcodes=false,render=false-4      375536344     342876828     -8.70%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=5,shortcodes=false,render=false-4      402182716     351249516     -12.66%
BenchmarkSiteBuilding/frontmatter=TOML,num_root_sections=1,num_pages=5000,tags_per_page=20,shortcodes=false,render=false-4     503704984     389290704     -22.71%
```

See #3541
Fixes #3464

/cc @bogem @spf13 @moorereason