@bogem 2017-05-31
Cool
I propose to use BurntSushi's lib in go-i18n library to have less dependencies and improve performance by parsing i18n configs. What do you think?