@bep 2017-05-31
>What do you think?

No, that is not a good idea.