@bep 2017-06-02
I'm slightly confused by some benchmarks here. My theory was that the taxonomy handling had a bottle neck in the form of a expensive inner loop or something (there may be improvements here as well), it seems that most of the time is wasted in front matter handling.

This is with go-toml:

```
▶ ./benchSite.sh "[YAML|TOML],num_pages=500,num_tags=.{3},tags_per_page=.{2}" | prettybench
Running with BenchmarkSiteBuilding/[YAML|TOML],num_pages=500,num_tags=.{3},tags_per_page=.{2}
PASS
benchmark                                                                  iter        time/iter       bytes alloc                allocs
---------                                                                  ----        ---------       -----------                ------
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=100,tags_per_page=20-4      1    4125.10 ms/op   1279495400 B/op    32880963 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=100,tags_per_page=50-4      1    9910.06 ms/op   3165169816 B/op    81733203 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=100,tags_per_page=80-4      1   15684.43 ms/op   5123012552 B/op   130582502 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=500,tags_per_page=20-4      1    3857.51 ms/op   1250655152 B/op    32902891 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=500,tags_per_page=50-4      1   10270.80 ms/op   3131353728 B/op    81752054 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=500,tags_per_page=80-4      1   17818.15 ms/op   5058262784 B/op   130602175 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=100,tags_per_page=20-4     10     125.82 ms/op     50894960 B/op      893454 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=100,tags_per_page=50-4     10     179.97 ms/op     70418993 B/op     1697637 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=100,tags_per_page=80-4      5     260.66 ms/op     89965144 B/op     2504681 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=500,tags_per_page=20-4     10     115.66 ms/op     53144384 B/op      942353 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=500,tags_per_page=50-4     10     180.49 ms/op     74482152 B/op     1796451 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=500,tags_per_page=80-4      5     265.71 ms/op     96553563 B/op     2646091 allocs/op
```

This is with BurntSushi (i.e. this branch):

```
▶ ./benchSite.sh "[YAML|TOML],num_pages=500,num_tags=.{3},tags_per_page=.{2}" | prettybench
Running with BenchmarkSiteBuilding/[YAML|TOML],num_pages=500,num_tags=.{3},tags_per_page=.{2}
PASS
benchmark                                                                  iter        time/iter       bytes alloc                allocs
---------                                                                  ----        ---------       -----------                ------
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=100,tags_per_page=20-4      1    4004.11 ms/op   1279493168 B/op    32880948 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=100,tags_per_page=50-4      1    9645.26 ms/op   3165173432 B/op    81733210 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=100,tags_per_page=80-4      1   15084.10 ms/op   5122482416 B/op   130582469 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=500,tags_per_page=20-4      1    3870.11 ms/op   1251179360 B/op    32902874 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=500,tags_per_page=50-4      1    9621.80 ms/op   3131361296 B/op    81752091 allocs/op
BenchmarkSiteBuilding/YAML,num_pages=500,num_tags=500,tags_per_page=80-4      1   15217.37 ms/op   5058251232 B/op   130602142 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=100,tags_per_page=20-4     20      71.73 ms/op     38847736 B/op      439204 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=100,tags_per_page=50-4     20      91.51 ms/op     45141734 B/op      636467 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=100,tags_per_page=80-4     10     109.01 ms/op     51168075 B/op      832525 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=500,tags_per_page=20-4     20      79.27 ms/op     41046766 B/op      461352 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=500,tags_per_page=50-4     10     102.60 ms/op     48267592 B/op      668635 allocs/op
BenchmarkSiteBuilding/TOML,num_pages=500,num_tags=500,tags_per_page=80-4     10     121.80 ms/op     56941638 B/op      875482 allocs/op
ok  	github.com/spf13/hugo/hugolib	72.310s
````
And while BurntSushi is obviously faster, there is some fishy going on in YAML with front matter ala:

```yaml
tags:
- tag1
- tag2
````

I will create a more focused benchmark outside of this PR. I'm not sure this is worth merging anymore. Will revisit this in a minute...